new util function for jwt
